requirements.txt
shortuuid
aws-cdk-lib
constructs
